+++ src/main/java/com/sunny/backend/comment/domain/Comment.java
@@ -1,7 +1,8 @@
-package com.sunny.backend.entity;
+package com.sunny.backend.comment.domain;
 
 
 import com.sunny.backend.community.domain.Community;
+import com.sunny.backend.entity.BaseTime;
 import com.sunny.backend.user.Users;
 import javax.validation.constraints.NotBlank;
 import lombok.*;
