+++ src/main/java/com/sunny/backend/comment/repository/CommentCustomRepository.java
@@ -1,7 +1,7 @@
-package com.sunny.backend.repository.comment;
+package com.sunny.backend.comment.repository;
 
 import com.sunny.backend.dto.response.comment.CommentResponse;
-import com.sunny.backend.entity.Comment;
+import com.sunny.backend.comment.domain.Comment;
 
 import java.util.List;
 import java.util.Optional;
