+++ src/main/java/com/sunny/backend/comment/repository/CommentRepository.java
@@ -1,8 +1,6 @@
-package com.sunny.backend.repository.comment;
+package com.sunny.backend.comment.repository;
 
-import com.sunny.backend.dto.response.comment.CommentResponse;
-import com.sunny.backend.dto.response.community.CommunityResponse;
-import com.sunny.backend.entity.Comment;
+import com.sunny.backend.comment.domain.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
