+++ src/main/java/com/sunny/backend/comment/service/CommentService.java
@@ -1,11 +1,10 @@
-package com.sunny.backend.service.comment;
+package com.sunny.backend.comment.service;
 
 import static com.sunny.backend.common.CommonErrorCode.*;
 
 import com.sunny.backend.community.repository.CommunityRepository;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -18,9 +17,9 @@ import com.sunny.backend.common.ResponseService;
 import com.sunny.backend.dto.request.comment.CommentRequest;
 import com.sunny.backend.dto.request.comment.CommentRequestMapper;
 import com.sunny.backend.dto.response.comment.CommentResponse;
-import com.sunny.backend.entity.Comment;
+import com.sunny.backend.comment.domain.Comment;
 
-import com.sunny.backend.repository.comment.CommentRepository;
+import com.sunny.backend.comment.repository.CommentRepository;
 import com.sunny.backend.security.userinfo.CustomUserPrincipal;
 import com.sunny.backend.user.Users;
 import com.sunny.backend.community.domain.Community;
@@ -55,7 +54,7 @@ public class CommentService {
 		commentResponse.setChildren(comment.getChildren()
 				.stream()
 				.map(childComment -> mapCommentToResponse(childComment, currentUser))
-				.collect(Collectors.toList())
+				.toList()
 		);
 		return commentResponse;
 	}
@@ -69,7 +68,7 @@ public class CommentService {
 		List<CommentResponse> commentResponses = comments.stream()
 				.filter(comment -> comment.getParent() == null)
 				.map(comment -> mapCommentToResponse(comment, user))
-				.collect(Collectors.toList());
+				.toList();
 		return responseService.getListResponse(HttpStatus.OK.value(), commentResponses, "댓글을 조회했습니다.");
 	}
 	//댓글 등록
@@ -86,15 +85,11 @@ public class CommentService {
 		if (commentRequestDTO.getParentId() != null) {
 			parentComment = commentRepository.findById(commentRequestDTO.getParentId())
 					.orElseThrow(() -> new CommonCustomException(COMMENT_NOT_FOUND));
-
 			if (parentComment.getParent() != null) {
 				throw new CommonCustomException(REPLYING_NOT_ALLOWED);
 			}
 			comment.setParent(parentComment);
 		}
-
-		//To do : setter 제외하고 도메인에서 함수로 처리
-
 		comment.setCommunity(community);
 		comment.setContent(commentRequestDTO.getContent());
 		comment.setUsers(user);
@@ -102,41 +97,33 @@ public class CommentService {
 		boolean isPrivate = commentRequestDTO.getIsPrivated();
 		comment.setIsPrivated(isPrivate);
 
-		Comment saveComment = commentRepository.save(comment);
-		if (user.getCommentList() == null) {
-			user.addComment(comment);
-		}
-
+		user.addComment(comment);
 		return responseService.getSingleResponse(HttpStatus.OK.value(),
 				new CommentResponse(comment.getId(), comment.getUsers().getName(), comment.getContent(),
 						comment.getCreatedDate(), comment.getUpdatedDate()), "댓글을 등록했습니다.");
 
 	}
 
-	//댓글 삭제
 	@Transactional
 	public ResponseEntity<CommonResponse.GeneralResponse> deleteComment(
 			CustomUserPrincipal customUserPrincipal, Long commentId) {
 		Comment comment = commentRepository.findCommentByIdWithParent(commentId)
 				.orElseThrow(() -> new CommonCustomException(COMMENT_NOT_FOUND));
 		if (checkCommentLoginUser(customUserPrincipal, comment)) {
-			if (comment.getChildren().size() != 0) { // 자식이 있으면 상태만 변경
+			if (comment.getChildren().size() != 0) {
 				comment.changeIsDeleted(true);
-			} else { // 삭제 가능한 조상 댓글을 구해서 삭제
+			} else {
 				commentRepository.delete(getDeletableAncestorComment(comment));
 			}
 		}
 		return responseService.getGeneralResponse(HttpStatus.OK.value(), "댓글을 삭제 하였습니다.");
 	}
 
-
-	//조상 댓글 있는지 check
 	private Comment getDeletableAncestorComment(Comment comment) {
 		Comment parent = comment.getParent(); // 현재 댓글의 부모를 구함
 		if (parent != null && parent.getChildren().size() == 1 && parent.getIsDeleted())
-			// 부모가 있고, 부모의 자식이 1개(지금 삭제하는 댓글)이고, 부모의 삭제 상태가 TRUE인 댓글이라면 재귀
 			return getDeletableAncestorComment(parent);
-		return comment; // 삭제해야하는 댓글 반환
+		return comment;
 	}
 
 	@Transactional
@@ -156,7 +143,6 @@ public class CommentService {
 						comment.getCreatedDate(), comment.getUpdatedDate()), "댓글을 수정했습니다.");
 	}
 
-	//수정 및 삭제 권한 체크
 	private boolean checkCommentLoginUser(CustomUserPrincipal customUserPrincipal, Comment comment) {
 		if (!Objects.equals(comment.getUsers().getName(), customUserPrincipal.getName())) {
 			return false;
