+++ src/main/java/com/sunny/backend/dto/request/comment/CommentRequestMapper.java
@@ -2,7 +2,7 @@ package com.sunny.backend.dto.request.comment;
 
 
 import com.sunny.backend.common.GenericMapper;
-import com.sunny.backend.entity.Comment;
+import com.sunny.backend.comment.domain.Comment;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
