+++ src/main/java/com/sunny/backend/dto/response/comment/CommentResponse.java
@@ -1,7 +1,7 @@
 package com.sunny.backend.dto.response.comment;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.sunny.backend.entity.Comment;
+import com.sunny.backend.comment.domain.Comment;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
